libvtpm: Add libcrt
The libcrt provides a static library (libcrt.a) with all the libc symbols
required to build OpenSSL and the Microsoft TPM 2.0; it is a subset of the musl
libc. In order to keep it small and simple, we organized its code similar to
OVMF; the library headers are just a wrapper to libcrt.h, which contains only
the definitions we need. The C files are organized by namespace, but all the
stubs are centralized in the file stub.c.

The required malloc(), realloc(), calloc(), free(), abort() and serial_out()
will be implemented in Rust.

The required rand() function uses the RDRAND instruction to get random bytes.

The time() function uses the RDTSC instruction to get the current time.

Co-developed-by: Vikram Narayanan <vikram186@gmail.com>
Signed-off-by: Vikram Narayanan <vikram186@gmail.com>
Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>