libvtpm/Makefile: Build libcrypto.a
Build the libcrypto.a from OpenSSL and include it into the libvtpm.a.

As of now, the upstream Microsoft TPM 2.0 requires OpenSSL 1.1.1d or higher,
but OpenSSL 3.x support is under review in a pull request.

This patch builds OpenSSL 1.1.1q similar to how OVMF does it, although OVMF has
moved to OpenSSL 3; we build it providing a SVSM configuration and disabling
crypto modules that are not needed. With that, the libcrypto.a generated is
optimized for memory footprint and that also reduces the build complexity.

Later we can move to OpenSSL 3.x.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>