libvtpm/Makefile: Build libcrypto.a
Build the libcrypto.a from OpenSSL and include it into the libvtpm.a.

As of now, the upstream Microsoft TPM 2.0 requires OpenSSL 1.1.1d or higher.
OpenSSL 3.x support is still under review in a pull request.

We build OpenSSL 1.1.1q similar to how OVMF used to do it by providing a SVSM
configuration and disabling crypto modules that are not needed.

Later we can move to OpenSSL 3.x.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>