libvtpm/Makefile: Build libtpm.a and libplatform.a
Build the libtpm.a and libplatform.a from the Microsoft TPM 2.0 and include
them into the libvtpm.a.

Due to yet unknown reasons, the TPM Self Test command is failing causing all
subsequent commands to fail. To workaround that we are disabling the self tests
(-DSELF_TEST=NO) for now.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>