libvtpm/Makefile: Build libtpm.a and libplatform.a
Build the libtpm.a and libplatform.a from the Microsoft TPM 2.0 and include
them into the libvtpm.a.

We are disabling the TPM SelfTests in the Microsoft TPM 2.0. Both the OVMF and
the guest kernel TPM drivers call the TPM SelfTest in the probe(), but the
ECDSA tests are failing reporting that the point is at inifinity; consequently,
the TPM enters in failure mode causing all subsequent TPM commands to also
fail. As a workaround, to unblock development, we are disabling the SelfTests
(-DSELF_TEST=NO), forcing the SelfTests to always return success.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>