libvtpm: Build and link the libvtpm.a
The libvtpm.a is linked to the stage1/kernel.elf only when the vtpm feature is
enabled. Ideally, the libvtpm.a build should also follow the vtpm feature. For
now we are not doing that because the Makefile is not a good/natural place to
track changes to the cargo default feature.

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>